+++ releasenotes.md
@@ -3,3 +3,4 @@ Whats New in v6.4
 - Fix supplier's details update modal box
 - Add available var list in SDS Html Editor
 - Added PV Online in ingredient search provider
+- Added option to import single ingredient data from pvonline